Refactor import statements to use destructuring for translateAttribute from handleLocales
